+++ backend/package.json
@@ -25,6 +25,7 @@
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.6.3",
     "multer": "^1.4.5-lts.1",
+    "nodemailer": "^6.9.16",
     "path": "^0.12.7"
   },
   "devDependencies": {
