+++ client/src/components/RatingList.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import Cookies from 'js-cookie';
 import { jwtDecode } from "jwt-decode";
-import "./RatingList.css"; 
+import ".//css/RatingList.css"; 
 
 const UserRatings = () => {
   const [ratings, setRatings] = useState([]);
