+++ client/src/components/css/RatingList.css
