+++ client/src/pages/Home.jsx
@@ -1,4 +1,4 @@
-import "./styles.css";
+import ".//css/styles.css";
 const Home = (props) => {
   return (
     <>
