+++ client/src/pages/StudentDashboard.jsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import Cookies from 'js-cookie';
 import { jwtDecode } from 'jwt-decode';
-import "./Dashboard.css";
+import ".//css/Dashboard.css";
 import TabContainer from "../components/TabContainer";
 import Button from "../components/Button";
 import GroupList from "../components/GroupList";
